+++ Sources/KrillCLI/KrillCLI.swift
@@ -36,4 +36,5 @@ struct Krill: AsyncParsableCommand {
         ],
         defaultSubcommand: RunCommand.self
     )
+
 }